Extract out coursework/exam ratio chart into its own file
