+++ frontend/src/pages/category-page.tsx
@@ -60,30 +60,7 @@ import { EuclidCodeBadge } from "../components/euclid-code-badge";
 import { useCategoryTabs } from "../hooks/useCategoryTabs";
 import CategoryStatsComponent from "../components/category-stats";
 import { useQuickSearchFilter } from "../components/Navbar/QuickSearch/QuickSearchFilterContext";
-import EChartsCore from "react-echarts-library/core";
-import * as echarts from "echarts/core";
-import type { EChartsOption } from "echarts";
-import { PieChart } from "echarts/charts";
-import { GridComponent, TitleComponent } from "echarts/components";
-import { SVGRenderer } from "echarts/renderers";
-
-echarts.use([PieChart, GridComponent, TitleComponent, SVGRenderer]);
-
-const pieChartOptions = (cw_exam_ratio: number[]): EChartsOption => ({
-  series: [
-    {
-      type: "pie",
-      data: cw_exam_ratio,
-      labelLine: { show: false },
-      silent: true,
-      radius: "100%",
-      color: [
-        "var(--mantine-primary-color-3)",
-        "var(--mantine-primary-color-9)",
-      ],
-    },
-  ],
-});
+import { CourseworkExamRatioChart } from "../components/Charts/CourseworkExamRatioChart";
 
 interface CategoryPageContentProps {
   onMetaDataChange: (newMetaData: CategoryMetaData) => void;
@@ -456,9 +433,8 @@ const CategoryPageContent: React.FC<CategoryPageContentProps> = ({
                               Semester {course.delivery_ordinal}
                             </Text>
                             <Group gap="xs">
-                              <EChartsCore
-                                echarts={echarts}
-                                option={pieChartOptions(course.cw_exam_ratio)}
+                              <CourseworkExamRatioChart
+                                cw_exam_ratio={course.cw_exam_ratio}
                                 style={{ height: 20, width: 20 }}
                               />
                               <Text>
